Added features for disabling and enabling popups
There is now a contextual menu option to enable Rikaitan. Also added an
event listener so that if the extension is active, there will be an
option to enable or disable popups. Also changed the hide popup key
from esc to h, and made the functionality visible to users in the mini
help menu.