Fixed entryData and sanseidoState
Cause of sensed definitions eventually breaking the entire extension
was caused by the extension incrementing sanseidoState to 3 as it
failed to find J-J definitions, but never reset itself to 0. Also fixed
a sequence in entryData checks that looked at the kana before the
kanji, causing some words to have the wrong definitions.